Split auto_utils into two files
